requirements.txt
flask
flask-cors
boto3
numpy
pandas
requests
shortuuid
pyjwt
uwsgi
python-socketio[client]
